@@ -1,23 +1,24 @@
-package su.spyme.moonhorse;
+package su.spyme.justhorse;
 
 import org.bukkit.entity.Horse;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.EntityDamageByEntityEvent;
+import org.bukkit.event.entity.EntityDamageEvent;
 import org.bukkit.event.entity.EntityDeathEvent;
 import org.bukkit.event.inventory.InventoryClickEvent;
 import org.bukkit.event.inventory.InventoryOpenEvent;
 import org.bukkit.event.player.PlayerQuitEvent;
 import org.bukkit.event.vehicle.VehicleExitEvent;
-import su.spyme.moonhorse.gui.GuiMenuHorse;
-import su.spyme.moonhorse.utils.GuiMenu;
+import su.spyme.justhorse.gui.GuiMenuHorse;
+import su.spyme.justhorse.utils.GuiMenu;
 
 import java.util.Map;
 
 public class Listeners implements Listener{
 
-    private Map<Player, Horse> horses = Main.horses;
+    private final Map<Player, Horse> horses = Main.horses;
 
     @EventHandler
     public void onGetOff(VehicleExitEvent event){
@@ -47,9 +48,9 @@ public class Listeners implements Listener{
 
     @EventHandler
     public void onDeath(EntityDeathEvent event){
-        if((event.getEntity() instanceof Horse) && Main.death){
-            event.setCancelled(true);
-            Horse horse = (Horse) event.getEntity();
+        if((event.getEntity() instanceof Horse horse) && Main.death){
+            event.getDrops().clear();
+            event.setDroppedExp(0);
             if(horse.getPassengers().isEmpty()) return;
             Player player = (Player) horse.getPassengers().get(0);
             Main.horses.remove(player);
@@ -62,8 +63,17 @@ public class Listeners implements Listener{
 
     @EventHandler
     public void onDamage(EntityDamageByEntityEvent event){
-        if((event.getEntity() instanceof Horse) && Main.damage){
-            Horse horse = (Horse) event.getEntity();
+        if((event.getEntity() instanceof Horse horse) && Main.damage){
+            Player player = (Player) horse.getPassengers().get(0);
+            if(horses.containsKey(player)){
+                event.setCancelled(true);
+            }
+        }
+    }
+
+    @EventHandler
+    public void onDamage(EntityDamageEvent event){
+        if((event.getEntity() instanceof Horse horse) && Main.damage){
             Player player = (Player) horse.getPassengers().get(0);
             if(horses.containsKey(player)){
                 event.setCancelled(true);
@@ -75,8 +85,8 @@ public class Listeners implements Listener{
     public void onClick(InventoryClickEvent event){
         Player player = (Player) event.getWhoClicked();
         if(player.getOpenInventory().getTitle().equals(Main.horse + " " + player.getName())){
-            event.setCancelled(true);
             if(Main.enableSettings){
+                event.setCancelled(true);
                 GuiMenu.openGui(new GuiMenuHorse(player));
             }
         }
