@@ -1,4 +1,4 @@
-package su.spyme.moonhorse.utils;
+package su.spyme.justhorse.utils;
 
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
@@ -22,7 +22,7 @@ import java.util.Set;
 public class GuiMenu implements Listener{
     private Player player; //Игрок
     private Inventory inventory; //Инвентарь
-    private Set<GuiItem> items = new HashSet<>(); //Предметы
+    private final Set<GuiItem> items = new HashSet<>(); //Предметы
 
     /**
      * Создание гуи
@@ -99,21 +99,12 @@ public class GuiMenu implements Listener{
     }
 
     private static int getWidth(InventoryType type){
-        switch(type){
-            case CHEST:
-            case PLAYER:
-            case ENDER_CHEST:
-            case SHULKER_BOX:
-                return 9;
-            case DISPENSER:
-            case DROPPER:
-            case WORKBENCH:
-                return 3;
-            case HOPPER:
-                return 5;
-            default:
-                return -1;
-        }
+        return switch(type){
+            case CHEST, PLAYER, ENDER_CHEST, SHULKER_BOX -> 9;
+            case DISPENSER, DROPPER, WORKBENCH -> 3;
+            case HOPPER -> 5;
+            default -> -1;
+        };
     }
 
     int getHeight(){
